Use monotonic time
When using xinet.d to limit rsync connections and time is set back,
the connection limit is reached very quickly and rsync gets deactivated.
If time is changed again, rsync is never reactivated.

$ date -s "xxx"
xinetd[xxx]: Deactivating service rsync due to excessive incoming connections. Restarting in 10 seconds.

The timer of xinet.d is based on time() function, and it is affected by system time.

Signed-off-by: Li Wang <li.wang@windriver.com>